Version:2.0 - Single Entry Point App/requirements.txt
python-dotenv
openai
requests
pelican
google-generativeai
vertexai